Fix tests
For some reasons, to run these tests successfully along with other tests, loading the module must be in its own beforeEach block. If not, module will be undefined. However, if we were to run these test in isolation, it would pass without using a separate beforeEach block to load the module. So strange!